requirements.txt
pyrogram
tgcrypto
python-dotenv
Pillow
reportlab
